requirements.txt
aiogram>=3.15.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
